feat: add new services

+++ scrapper/src/main/java/backend/academy/scrapper/service/LinkServiceImpl.java
@@ -1,14 +1,12 @@
 package backend.academy.scrapper.service;
 
-import backend.academy.scrapper.exceptions.NotExistLinkException;
 import backend.academy.scrapper.exceptions.NotExistTgChatException;
 import backend.academy.scrapper.exceptions.NotTrackLinkException;
 import backend.academy.scrapper.models.Link;
 import backend.academy.scrapper.models.LinkInfo;
-import backend.academy.scrapper.repository.LinksRepository;
+import backend.academy.scrapper.repository.database.LinksRepository;
 import java.time.OffsetDateTime;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,8 +17,11 @@ public class LinkServiceImpl implements LinkService {
 
     @Override
     public LinkInfo addLink(long chatId, Link link) {
-        linksRepository.findById(chatId)
-            .orElseThrow(NotExistTgChatException::new);
+        List<LinkInfo> linksById = linksRepository.findById(chatId);
+
+        if (linksById.isEmpty()){
+            throw new NotExistTgChatException();
+        }
 
         LinkInfo linkInfo = new LinkInfo();
         linkInfo.link(link);
@@ -31,8 +32,11 @@ public class LinkServiceImpl implements LinkService {
 
     @Override
     public LinkInfo removeLink(long chatId, String uri) {
-        linksRepository.findById(chatId)
-            .orElseThrow(NotExistTgChatException::new);
+        List<LinkInfo> linksById = linksRepository.findById(chatId);
+
+        if (linksById.isEmpty()){
+            throw new NotExistTgChatException();
+        }
 
         return linksRepository.deleteLink(chatId, uri)
             .orElseThrow(NotTrackLinkException::new);
@@ -40,8 +44,12 @@ public class LinkServiceImpl implements LinkService {
 
     @Override
     public List<LinkInfo> getLinks(long chatId) {
-        return linksRepository.findById(chatId)
-            .orElseThrow(NotExistTgChatException::new);
+        List<LinkInfo> linksById = linksRepository.findById(chatId);
+
+        if (linksById.isEmpty()){
+            throw new NotExistTgChatException();
+        }
+        return linksById;
     }
 
 }
